Fixgallery

+++ controllers/user.controller.js
@@ -7,21 +7,30 @@ require("dotenv").config();
 module.exports = {
   // post: /register
   register: (req, res) => {
-    const data = req.body;
-    // kurang 
+    let { name, email, password } = req.body;
+    let data = { name, email, password };
     try {
+      if (name.length == 0 || email.length <= 3 || password.length == 0) {
+        return res
+          .status(406)
+          .send({
+            message: "Data given doesn't meet the standards of required fields",
+          });
+      }
       const saltRounds = 10;
       const hash = bcrypt.hashSync(data.password, saltRounds);
       data.password = hash;
+      // console.log(data)
+
       const user = new User(data);
 
       user.save();
 
-      res.status(200).json({
-        message: "data has been created!!",
+      res.status(201).json({
+        message: "Sign Up success!",
       });
     } catch (error) {
-      res.status(500).send({
+      res.status(500).json({
         message: "server error",
         error: error.message,
       });
@@ -39,22 +48,25 @@ module.exports = {
         });
       }
       const checkPwd = bcrypt.compareSync(data.password, user.password);
-  
+
       if (checkPwd) {
         const token = jwt.sign({ user }, process.env.TOKEN_KEY, {
           expiresIn: 86400,
         }); //expires in 24 hours
-        res.header("x-access-token", token).status(200).json({
-          message: "Anda berhasil login",
-          token,
-        });
+        res
+          .header("x-access-token", token)
+          .status(200)
+          .json({
+            message: `Selamat Datang ${user.name.split(" ")[0]}`,
+            token: token,
+          });
       } else {
         res.status(400).json({
           message: "Login gagal",
         });
-      }  
+      }
     } catch (error) {
-      res.status(500).send({
+      res.status(500).json({
         message: "server error",
         error: error.message,
       });
@@ -64,14 +76,14 @@ module.exports = {
   // get:/
   getAllUser: async (req, res) => {
     try {
-      const user = await User.find({}, "-__v");
+      const user = await User.find({}, "-__v -createdAt -password");
       // console.log(req.user.user.email)
-      res.json({
+      res.status(200).json({
         message: "get all user success",
         data: user,
       });
     } catch (error) {
-      res.status(500).send({
+      res.status(500).json({
         message: "server error",
         error: error.message,
       });
@@ -84,29 +96,22 @@ module.exports = {
     try {
       if (!mongoose.Types.ObjectId.isValid(id))
         return res.status(400).json({ message: "invalid id" });
-      const user = await User.findById(id);
+      const user = await User.findById(
+        id,
+        "-__v -createdAt -password"
+      ).populate("kelas challenge", "-description -__v -materi -categories");
+      if (user === null){
+        return res.status(404).json({
+          message: "user dooesn't exist ",
+        });
+      }
       res.status(200).json({
-        message: "get a user succes",
+        message: `get ${user.email} success`,
         data: user,
       });
-    } catch (error) {
-      res.status(404).send({
-        message: "user doesn't exist",
-        error: error.message,
-      });
-    }
-  },
 
-  // update: /:id
-  updateUserByID: async (req, res) => {
-    const id = req.params;
-    try {
-      const data = req.body;
-      // if (data.)
-      const user = await User.findByIdAndUpdate(id, data);
-      console.table(user);
     } catch (error) {
-      res.status(500).send({
+      res.status(500).json({
         message: "server error",
         error: error.message,
       });
@@ -115,12 +120,19 @@ module.exports = {
 
   // type 2 update: /:id
   updateUserByID2: async (req, res) => {
-    const id = req.params;
+    const id = req.params.id;
     try {
       const data = req.body;
+      if (!mongoose.Types.ObjectId.isValid(id))
+        return res.status(400).json({ message: "invalid id" });
       // if (data.)
-      const user = await User.findById(id);
-      console.table(user);
+      const user = await User.findById(id, "-__v -createdAt");
+      if (user === null){
+        return res.status(404).json({
+          message: "user dooesn't exist ",
+        });
+      }
+      // console.log(user);
       if (data.name) {
         user.name = data.name;
       }
@@ -128,7 +140,9 @@ module.exports = {
         user.email = data.email;
       }
       if (data.password) {
-        user.password = data.password;
+        const saltRounds = 10;
+        const hash = bcrypt.hashSync(data.password, saltRounds);
+        user.password = hash;
       }
       if (data.role) {
         user.role = data.role;
@@ -139,16 +153,16 @@ module.exports = {
       if (data.tgl_lahir) {
         user.tgl_lahir = data.tgl_lahir;
       }
-      if (data.jns.kelamin) {
-        user.jns.kelamin = data.jns.kelamin;
+      if (data.jns_kelamin) {
+        user.jns_kelamin = data.jns_kelamin;
       }
       for (let item in data.kelas) {
-        if (data.kelas[item])
-          user.kelas[item] = data.kelas[item];
+        if (!user.kelas.includes(data.kelas[item]))
+          user.kelas.push(data.kelas[item]);
       }
       for (let item in data.challenge) {
-        if (data.challenge[item])
-          user.challenge[item] = data.challenge[item];
+        if (!user.challenge.includes(data.challenge[item]))
+          user.challenge.push(data.challenge[item]);
       }
       if (data.social_media.insta) {
         user.social_media.insta = data.social_media.insta;
@@ -159,42 +173,42 @@ module.exports = {
       if (data.social_media.other) {
         user.social_media.other = data.social_media.other;
       }
-      await user.save();
+
+      // console.log(user)
+      await User.findByIdAndUpdate(id, user);
+      // await user.save();
+
       res.status(200).json({
-        massage: "success",
+        massage: `user ${user.email} updated`,
         data: user,
       });
     } catch (error) {
-      res.status(500).send({
+      res.status(500).json({
         message: "server error",
         error: error.message,
       });
     }
   },
-
-  // _id
-  // name
-  // email
-  // password
-  // role
-  // sekolah
-  // tgl_lahir
-  // jns_kelamin
-  // kelas: complete: [] , progress: []
-  // challenge: complete: [] , progress: []
-  // social_media: insta: ,fb: , other:
+  // updatePasswordById:
 
   // delete: /:id
   deleteUserByID: async (req, res) => {
     const { id } = req.params;
     try {
+      if (!mongoose.Types.ObjectId.isValid(id))
+        return res.status(400).json({ message: "invalid id" });
       const user = await User.findByIdAndDelete(id);
+      if (user === null){
+        return res.status(404).json({
+          message: "user dooesn't exist ",
+        });
+      }
       res.status(200).json({
-        message: `user id ${id} succesfully deleted`,
+        message: `user ${user.email} deleted`,
       });
     } catch (error) {
-      res.status(404).send({
-        message: "user doesn't exist",
+      res.status(500).json({
+        message: `server error`,
         error: error.message,
       });
     }
